requirements.txt
duckdb==0.9.2
fastapi==0.109.0
fastparquet==2024.2.0
folium==0.15.1
h3==3.7.6
numpy==1.26.3
geopandas==0.14.2
imagecodecs==2024.1.1
requests==2.31.0
scipy==1.12.0
shapely==2.0.2
uvicorn==0.27.0.post1
pandas==2.2.0
PyYAML==6.0.1
pydantic==2.6.0
rasterio==1.3.9
requests==2.31.0
xarray==2023.1.0